+++ packages/components/src/assets/MagnifyingGlass.tsx
@@ -1,0 +1,39 @@
+import React from 'react'
+import type { FC } from 'react'
+
+// Icon from Phosphor Icons
+const MagnifyingGlass: FC = () => (
+  <svg
+    xmlns="http://www.w3.org/2000/svg"
+    fill="currentColor"
+    viewBox="0 0 256 256"
+    strokeWidth="16"
+    width={24}
+    height={24}
+  >
+    <rect width="256" height="256" fill="none"></rect>
+    <circle
+      cx="116"
+      cy="116"
+      r="84"
+      fill="none"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="16"
+    ></circle>
+    <line
+      x1="175.4"
+      y1="175.4"
+      x2="224"
+      y2="224"
+      fill="none"
+      stroke="currentColor"
+      strokeLinecap="round"
+      strokeLinejoin="round"
+      strokeWidth="16"
+    ></line>
+  </svg>
+)
+
+export default MagnifyingGlass