+++ packages/components/src/hooks/index.ts
@@ -1,3 +1,3 @@
-export { default as UIProvider, useUI } from './UIProvider'
+export { default as UIProvider, Toast as ToastProps, useUI } from './UIProvider'
 export { useFadeEffect } from './useFadeEffect'
 export { useTrapFocus } from './useTrapFocus'