+++ packages/components/src/molecules/SearchProducts/SearchProductItemContent.tsx
@@ -1,0 +1,51 @@
+import React, { forwardRef } from 'react'
+import { Price } from '../..'
+import type { PriceFormatter } from '../../atoms/Price/Price'
+
+interface Price {
+  value: number
+  listPrice: number
+  formatter: PriceFormatter
+}
+
+export type SearchProductItemContentProps = {
+  /**
+   * Specifies the product's title.
+   */
+  title: string
+  /**
+   * Specifies product's prices.
+   */
+  price: Price
+}
+
+const SearchProductItemContent = forwardRef<
+  HTMLElement,
+  SearchProductItemContentProps
+>(function SearchProductItemContent({ price, title, ...otherProps }, ref) {
+  return (
+    <section ref={ref} data-fs-search-product-item-content {...otherProps}>
+      <p data-fs-search-product-item-title>{title}</p>
+      <span data-fs-search-product-item-prices>
+        <Price
+          value={price?.listPrice ? price.listPrice : 0}
+          formatter={price?.formatter}
+          testId="list-price"
+          data-value={price?.listPrice}
+          variant="listing"
+          SRText="Original price:"
+        />
+        <Price
+          value={price?.value ? price.value : 0}
+          formatter={price?.formatter}
+          testId="price"
+          data-value={price?.value}
+          variant="spot"
+          SRText="Price:"
+        />
+      </span>
+    </section>
+  )
+})
+
+export default SearchProductItemContent