+++ packages/components/src/molecules/SearchProducts/index.ts
@@ -1,0 +1,11 @@
+export { default } from './SearchProducts'
+export type { SearchProductsProps } from './SearchProducts'
+
+export { default as SearchProductItem } from './SearchProductItem'
+export type { SearchProductItemProps } from './SearchProductItem'
+
+export { default as SearchProductItemImage } from './SearchProductItemImage'
+export type { SearchProductItemImageProps } from './SearchProductItemImage'
+
+export { default as SearchProductItemContent } from './SearchProductItemContent'
+export type { SearchProductItemContentProps } from './SearchProductItemContent'