+++ packages/core/src/components/ui/SkuSelector/Selectors.tsx
@@ -1,9 +1,17 @@
-import { useRouter } from 'next/router'
-import type { ChangeEvent, HTMLAttributes } from 'react'
+import { HTMLAttributes, useCallback, useMemo } from 'react'
 
-import SkuSelector from './SkuSelector'
-import { navigateToSku } from './skuVariants'
-import type { SkuVariantsByName } from './skuVariants'
+import { Image } from '../Image'
+import {
+  SkuSelector as UISkuSelector,
+  SkuSelectorProps,
+  SkuOption,
+} from '@faststore/ui'
+import NextLink from 'next/link'
+
+export type SkuVariantsByName = Record<
+  string,
+  Array<{ alt: string; label: string; value: string; src?: string }>
+>
 
 interface Props extends HTMLAttributes<HTMLDivElement> {
   /**
@@ -20,64 +28,42 @@ interface Props extends HTMLAttributes<HTMLDivElement> {
   activeVariations: Record<string, string>
 }
 
-/**
- * Name of the property that's considered **dominant**. Which means that all
- * other varying properties will be filtered according to the current value
- * of this property.
- *
- * Ex: If `Red` is the current value for the 'Color' variation, we'll only
- * render possible values for 'Size' that are available in `Red`.
- */
-const DOMINANT_SKU_SELECTOR_PROPERTY = 'Color'
+const ImageComponent: SkuSelectorProps['ImageComponent'] = ({
+  src,
+  alt,
+  ...otherProps
+}) => (
+  <Image
+    src={src}
+    alt={alt}
+    width={20}
+    height={20}
+    loading="lazy"
+    {...otherProps}
+  />
+)
 
 function Selectors({
   slugsMap,
-  availableVariations,
   activeVariations,
+  availableVariations,
   ...otherProps
 }: Props) {
-  const router = useRouter()
-
-  // 'Color' variants are singled-out here because they will always be rendered
-  // as 'image' variants. And they're also the 'dominant' variants in our store.
-  const { Color: colorOptions, ...otherSkuVariants } = availableVariations
-
-  function handleOnChange(
-    e: ChangeEvent<HTMLInputElement>,
-    updatedVariationName: string
-  ) {
-    const newVariationValue = e.currentTarget.value
-
-    navigateToSku({
-      router,
-      slugsMap,
-      updatedVariationName,
-      selectorsState: activeVariations,
-      updatedVariationValue: newVariationValue,
-      dominantSku: DOMINANT_SKU_SELECTOR_PROPERTY,
-    })
-  }
-
   return (
     <section {...otherProps}>
-      {colorOptions && (
-        <SkuSelector
-          label="Color"
-          variant="image"
-          options={colorOptions}
-          activeValue={activeVariations.Color}
-          onChange={(e) => handleOnChange(e, 'Color')}
-        />
-      )}
-      {otherSkuVariants &&
-        Object.keys(otherSkuVariants).map((skuVariant) => (
-          <SkuSelector
-            variant="label"
+      {availableVariations &&
+        Object.keys(availableVariations).map((skuVariant) => (
+          <UISkuSelector
             key={skuVariant}
-            label={skuVariant}
-            options={otherSkuVariants[skuVariant]}
-            activeValue={activeVariations[skuVariant]}
-            onChange={(e) => handleOnChange(e, skuVariant)}
+            skuPropertyName={skuVariant}
+            availableVariations={availableVariations}
+            ImageComponent={ImageComponent}
+            activeVariations={activeVariations}
+            slugsMap={slugsMap}
+            linkProps={{
+              as: NextLink,
+              legacyBehavior: false,
+            }}
           />
         ))}
     </section>